Secure_storage: fix size type into uint32_t
Type size_t is a deprecated type used in GPD TEE Client API v1.0, since v1.1
size arguments are typed uint32_t. Secure storage examples was not updated.

This error was reported with NXP LSDK (flexbbuild system with
gcc-aach64-linux-gnu <aarch64-linux-gnu-gcc (Ubuntu/Linaro 7.3.0-27ubuntu1~18.04) 7.3.0) with trace:

secure_storage_ta.c:203:6: error: passing argument 4 of ‘TEE_ReadObjectData’ from incompatible pointer type [-Werror=incompatible-pointer-types]
&read_bytes);

+++ secure_storage/ta/secure_storage_ta.c
@@ -146,7 +146,7 @@ static TEE_Result read_raw_object(uint32_t param_types, TEE_Param params[4])
 	TEE_ObjectHandle object;
 	TEE_ObjectInfo object_info;
 	TEE_Result res;
-	size_t read_bytes;
+	uint32_t read_bytes;
 	char *obj_id;
 	size_t obj_id_sz;
 	char *data;
@@ -202,7 +202,7 @@ static TEE_Result read_raw_object(uint32_t param_types, TEE_Param params[4])
 	res = TEE_ReadObjectData(object, data, object_info.dataSize,
 				 &read_bytes);
 	if (res != TEE_SUCCESS || read_bytes != object_info.dataSize) {
-		EMSG("TEE_ReadObjectData failed 0x%08x, read %u over %u",
+		EMSG("TEE_ReadObjectData failed 0x%08x, read %" PRIu32 " over %u",
 				res, read_bytes, object_info.dataSize);
 		goto exit;
 	}
